refactor: move components related to pages back to their Pages module
