+++ packages/client/src/components/animate.jsx
@@ -43,18 +43,21 @@ const addmove = (c1, x1, x2, y1, y2, attributeNameX, attributeNameY) => {
 const Animation = {
   createTree: (dataSetToUse, svgRef) => {
     const width = svgRef.current.clientWidth;
-    //console.log(dataSetToUse)
     const svg = d3.select(svgRef.current);
     svg.selectAll("*").remove();
     const my = 60;
+
+    // draw links
     const p = svg
-      .append("g")
+      .append("g")              // "g" is used to group SVG shape together
       .attr("stroke", "black")
       .attr("stroke-width", "1")
       .selectAll("line")
       .data(dataSetToUse)
       .enter()
       .append("line")
+
+      // what's the difference between this and finalTree
       .attr("id", (c) => "l" + c.index)
       .attr("x1", (c) => (c.index === 1 ? null : getx(c.index, width)))
       .attr("y1", (c) => (c.index === 1 ? null : my * (getdepth(c.index) + 1)))
@@ -62,6 +65,8 @@ const Animation = {
         c.index === 1 ? null : getx(Math.floor(c.index / 2), width),
       )
       .attr("y2", (c) => (c.index === 1 ? null : my * getdepth(c.index)));
+
+    // draw node
     const c = svg
       .append("g")
       .attr("stroke", "black")
@@ -75,6 +80,8 @@ const Animation = {
       .attr("cx", (c) => getx(c.index, width))
       .attr("cy", (c) => my * (getdepth(c.index) + 1))
       .attr("r", 20);
+    
+    // draw label
     const t = svg
       .append("g")
       .attr("stroke", "black")
@@ -89,12 +96,14 @@ const Animation = {
       .attr("dy", (c) => my * (getdepth(c.index) + 1) + 5)
       .text((t) => t.value);
   },
-  fianlTree: (dataSetToUse, svgRef) => {
+
+  finalTree: (dataSetToUse, svgRef) => {
     const width = svgRef.current.clientWidth;
     const svg = d3.select(svgRef.current);
-    //console.log(dataSetToUse)
     svg.selectAll("*").remove();
     const my = 60;
+
+    // draw links
     const p = svg
       .append("g")
       .attr("stroke", "black")
@@ -103,6 +112,8 @@ const Animation = {
       .data(dataSetToUse)
       .enter()
       .append("line")
+
+      // what's the difference between this and createTree
       .attr("id", (c) => "l" + c.index)
       .attr("x1", (c, i) => (i === 0 ? null : getx(i + 1, width)))
       .attr("y1", (c, i) => (i === 0 ? null : my * (getdepth(i + 1) + 1)))
@@ -110,6 +121,9 @@ const Animation = {
         i === 0 ? null : getx(Math.floor((i + 1) / 2), width),
       )
       .attr("y2", (c, i) => (i === 0 ? null : my * getdepth(i + 1)));
+      //
+    
+    // draw node
     const c = svg
       .append("g")
       .attr("stroke", "black")
@@ -123,6 +137,8 @@ const Animation = {
       .attr("cx", (c, i) => getx(i + 1, width))
       .attr("cy", (c, i) => my * (getdepth(i + 1) + 1))
       .attr("r", 20);
+
+    // draw label
     const t = svg
       .append("g")
       .attr("stroke", "black")
