+++ packages/client/src/components/solver/algorithm.js
