+++ packages/client/src/pages/Algorithm/DijkstraComponent/DijkstraGraph.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import * as d3 from "d3";
 import { Button, Stack } from "@mui/material";
-import Graph from "./algorithm.js";
+import Graph from "../../../components/solver/algorithm.js";
 
 export const DijkstraGraph = () => {
   const [nodes, setNodes] = useState([
